hotfix: refactor user ID and pagination types from int to int64/int32

+++ internal/application/delete_user.go
@@ -7,7 +7,7 @@ import (
 )
 
 type DeleteUserUseCase interface {
-	Execute(ctx context.Context, id int) error
+	Execute(ctx context.Context, id int64) error
 }
 
 type deleteUserUseCase struct {
@@ -18,7 +18,7 @@ func NewDeleteUserUseCase(userRepository domain.UserRepository) DeleteUserUseCas
 	return &deleteUserUseCase{userRepository: userRepository}
 }
 
-func (d *deleteUserUseCase) Execute(ctx context.Context, id int) error {
+func (d *deleteUserUseCase) Execute(ctx context.Context, id int64) error {
 	if err := domain.ValidateAssignedUserID(id); err != nil {
 		return err
 	}
