+++ mocks/application/mock_UpdateUserEmailUseCase.go
@@ -23,7 +23,7 @@ func (_m *MockUpdateUserEmailUseCase) EXPECT() *MockUpdateUserEmailUseCase_Expec
 }
 
 // Execute provides a mock function with given fields: ctx, id, email
-func (_m *MockUpdateUserEmailUseCase) Execute(ctx context.Context, id int, email *domain.Email) error {
+func (_m *MockUpdateUserEmailUseCase) Execute(ctx context.Context, id int64, email *domain.Email) error {
 	ret := _m.Called(ctx, id, email)
 
 	if len(ret) == 0 {
@@ -31,7 +31,7 @@ func (_m *MockUpdateUserEmailUseCase) Execute(ctx context.Context, id int, email
 	}
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, int, *domain.Email) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, int64, *domain.Email) error); ok {
 		r0 = rf(ctx, id, email)
 	} else {
 		r0 = ret.Error(0)
@@ -47,15 +47,15 @@ type MockUpdateUserEmailUseCase_Execute_Call struct {
 
 // Execute is a helper method to define mock.On call
 //   - ctx context.Context
-//   - id int
+//   - id int64
 //   - email *domain.Email
 func (_e *MockUpdateUserEmailUseCase_Expecter) Execute(ctx interface{}, id interface{}, email interface{}) *MockUpdateUserEmailUseCase_Execute_Call {
 	return &MockUpdateUserEmailUseCase_Execute_Call{Call: _e.mock.On("Execute", ctx, id, email)}
 }
 
-func (_c *MockUpdateUserEmailUseCase_Execute_Call) Run(run func(ctx context.Context, id int, email *domain.Email)) *MockUpdateUserEmailUseCase_Execute_Call {
+func (_c *MockUpdateUserEmailUseCase_Execute_Call) Run(run func(ctx context.Context, id int64, email *domain.Email)) *MockUpdateUserEmailUseCase_Execute_Call {
 	_c.Call.Run(func(args mock.Arguments) {
-		run(args[0].(context.Context), args[1].(int), args[2].(*domain.Email))
+		run(args[0].(context.Context), args[1].(int64), args[2].(*domain.Email))
 	})
 	return _c
 }
@@ -65,7 +65,7 @@ func (_c *MockUpdateUserEmailUseCase_Execute_Call) Return(_a0 error) *MockUpdate
 	return _c
 }
 
-func (_c *MockUpdateUserEmailUseCase_Execute_Call) RunAndReturn(run func(context.Context, int, *domain.Email) error) *MockUpdateUserEmailUseCase_Execute_Call {
+func (_c *MockUpdateUserEmailUseCase_Execute_Call) RunAndReturn(run func(context.Context, int64, *domain.Email) error) *MockUpdateUserEmailUseCase_Execute_Call {
 	_c.Call.Return(run)
 	return _c
 }
